test: add underscore prefix to jest tests to avoid issues with Ava
